Remove wrong assertion of `offset`
When choosing n vertices from n+1 vertices, one vertex is excluded.
`offset` replesents this exlusion.

- - -

# Choose a[4] from A[4+1].

`offset` is used as follows.

## Exclude A[2]
(offset = 0) a[0] = A[0]
(offset = 0) a[1] = A[1]
(offset = 1) a[2] = A[2+1] = A[3]
(offset = 1) a[3] = A[3+1] = A[4]

## Exclude A[4]
(offset = 0) a[0] = A[0]
(offset = 0) a[1] = A[1]
(offset = 0) a[2] = A[2]
(offset = 0) a[3] = A[3]

- - -

As confirmed above, the final offset value can be either 0 or 1.
Old implementation with BUG, force final offset value is not 0.
Final `offset` value cannot be more than 2, because all vertices are different.
So, remove `offset` final value check from the code.